Refactor SearchPage
